+++ global.go
@@ -4,6 +4,6 @@ import "github.com/anaminus/rbxmk/rtypes"
 
 // Global contains values that available across an entire World.
 type Global struct {
-	Desc       *rtypes.RootDesc
+	Desc       *rtypes.Desc
 	AttrConfig *rtypes.AttrConfig
 }